Fix to issue reported from twitter user C__Kronenberg
Error occurred when the user above had a value of `.a` in a variable.  The logic for dropping cases with missing values fails to test for any missing value and only specifically tests for the missing value `.`.  Using the `missing()` function instead of the equality should take care of the issue.